Troca float por double no complexo

+++ Complexo.cpp
@@ -1,35 +1,35 @@
 #include "Complexo.h"
 
-Complexo::Complexo(float _real, float _imaginario) {
+Complexo::Complexo(double _real, double _imaginario) {
 	real = _real;
 	imaginario = _imaginario;
 }
 
 Complexo::~Complexo() {}
 
-float Complexo::getReal() {
+double Complexo::getReal() {
 	return real;
 }
 
-float Complexo::getImaginario() {
+double Complexo::getImaginario() {
 	return imaginario;
 }
 
 Complexo Complexo::operator+(Complexo n) {
-	float parte_real = real + n.getReal();
-	float parte_imaginaria = imaginario + n.getImaginario();
+	double parte_real = real + n.getReal();
+	double parte_imaginaria = imaginario + n.getImaginario();
 	return Complexo(parte_real, parte_imaginaria);
 }
 
 Complexo Complexo::operator-(Complexo n) {
-	float parte_real = real - n.getReal();
-	float parte_imaginaria = imaginario - n.getImaginario();
+	double parte_real = real - n.getReal();
+	double parte_imaginaria = imaginario - n.getImaginario();
 	return Complexo(parte_real, parte_imaginaria);
 }
 
 Complexo Complexo::operator*(Complexo n){
-	float r = (real * n.getReal() - imaginario * n.getImaginario());
-	float i = (real * n.getImaginario() + imaginario * n.getReal());
+	double r = (real * n.getReal() - imaginario * n.getImaginario());
+	double i = (real * n.getImaginario() + imaginario * n.getReal());
 	return Complexo(r, i);
 }
